move test to ip address suite
